Added New Notebook and Utils File
Significant change to the notebook design in a new file.

Now we are using pandas Data-Frame instead of NumPy 2D-Arrays.

Added Sections for every step in the project.

Made a utils.py file to contain all our preprocessing and helping function so we don't write it in our notebook.

Removed the dataset and put it in a new folder for direct loading of the datasets.